Remove `{to,from}Letter` from public API of `Location`.
This commit removes functions `{to,from}Letter` from the public API of
the `Location` module. To achieve this, we create a new internal module
`Location.Internal` to house the definitions of the `Location` type and
its related functions. We then selectively re-export just what we need
to expose from public module `Location`.